chore(Package.json): Add nps/nps-utils and revamp initialization
I honestly want to be able to keep this up to date, and the
way the postinstall was set up was a bit poop.

So now theres an explicit initialization step that needs running,
apart from that I've also added nps (I need to start maintaining
that as well...), since I really really honest-to-god like that
library.